Fix race during startup
When using netlink and the input layer, acpid opens the existing devices
in /dev/input/event* first and only after sets an inotify to be told
about newly created devices. If an event device is created between the
time acpid reads the directory and the time it sets the inotify, it
won't open that device and events provided by that device will be lost.

On my system, this race happens frequently for /dev/input/event8 "Video
Bus" during boot, preventing acpid from handling my system's brightness
keys.

The attached patch changes the ordering, setting the inotify before
opening the existing devices. It also adds a check to prevent
double-opening of a device if it is created between the inotify and the
reading of the directory.

From anomie.
Bug #18.
Debian bug #933230.