inotify: Also consider attribute changes for input devices
When acpid is running as a regular user it depends on the udev rules
for permitting access to the input event devices. However, when the
creation of the device node (IN_CREATE) is processed the processing
of the udev rules has not happened yet. Only a few millisecs later
the ownership and permissions of the device node has changed.

Adding support for inotify's IN_ATTRIB events to check if the input
device is already known. If not, just try to open it again.

This fixes a problem that input events from devices added during the
runtime (such as USB keyboards or USB headsets with volume control)
were not processed.

This belongs to issue CLIENT-24118.